调整一些参数

+++ utils/utils.py
@@ -28,7 +28,8 @@ def labels_to_string(label, vocabulary, blank_index=0):
 
 # 使用模糊删除方式删除文件
 def fuzzy_delete(dir, fuzzy_str):
-    for file in os.listdir(dir):
-        if fuzzy_str in file:
-            path = os.path.join(dir, file)
-            os.remove(path)
+    if os.path.exists(dir):
+        for file in os.listdir(dir):
+            if fuzzy_str in file:
+                path = os.path.join(dir, file)
+                os.remove(path)
